refactor: #7 move unit test to test/unit dir
